Feat: Finish client-server interaction through the RMI registry
Registered the promotion object in the server file, and created a client that uses said remote object and does tests with all of the available object functions.
Also created two .bat files to run the client and server without needing to open an IDE.
The RMIRegistry stills need to be run manually from the command line, but the startServer.bat already handles that.